requirements.txt
Flask==3.1.0
geopy==2.4.1
openmeteo_requests==1.3.0
pandas==2.2.3
pytz==2024.2
PyYAML==6.0.2
PyYAML==6.0.2
requests_cache==1.2.1
retry_requests==2.0.0
timezonefinder==6.5.8
